requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
celery[redis]
# Core libraries for FastAPI and async tasks

# Authentication & Security
authlib==1.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0

# External APIs
google-generativeai==0.3.0
python-quickbooks==0.9.1

# Document Processing
pdf2image==1.16.3
Pillow==10.1.0
boto3==1.29.7

# Data Processing
pandas==2.1.3
aiohttp==3.9.1

# Letter Generation
jinja2==3.1.2
weasyprint==60.1

# Database & Caching
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Type stubs
types-redis==4.6.0.11
boto3-stubs[s3]==1.29.7
types-requests==2.31.0.10

# Development dependencies
pre-commit
